move: PostDetailPage와 위젯 이동
